Implement node FSM code & remove 'reserving' state
The 'reserving' state wasn't really necessary, so I removed it for
simplicity. Nodes now go from 'requested' to 'starting' to 'active'
(assuming no errors occur).

This commit also introduces a bunch of new FSM code, and ties it into
the existing scheduler code. The FSM is now responsible for keeping the
node state updated properly in the metadata.